Previously, an incorrect defer was causing the statistics generation to increment every time a statistics was looked up from the cache. This would cause the staleness optimization to kick in an inconsistent manner and be disabled needlessly. To address this, this patch only increments the generation upon deletion.

After this patch we see:
```
name                                  old time/op    new time/op    delta
Sysbench/SQL/3node/oltp_read_only-10    2.97ms ±14%    2.69ms ± 7%  -9.43%  (p=0.008 n=9+10)

name                                  old alloc/op   new alloc/op   delta
Sysbench/SQL/3node/oltp_read_only-10    1.14MB ± 1%    1.13MB ± 0%  -1.11%  (p=0.000 n=9+9)

name                                  old allocs/op  new allocs/op  delta
Sysbench/SQL/3node/oltp_read_only-10     4.42k ± 2%     4.31k ± 0%  -2.55%  (p=0.000 n=9+9)
```

Fixes: #139876

Release note: None