@fqazi 2025-01-27
I had to add one more fix here to the lease generation stuff, since TestRenameTable started failing. For long running queries fetching the latest generation didn't work correctly, since those transactions observe the past. So, the second commit makes it so that the lease generation is cached at the start of any txn. An example of this:

```
    A simple example of this scenario:
    conn 1:
    BEGIN;
    SELECT * FROM t;

    conn 2:
    ALTER TABLE t RENAME TO t2;

    conn 1:
    SELECT * FROM t;
    COMMIT;
    SELECT * FROM t; -- the memo staleness check should fail.
```
@rafiss @mgartner Can you look at the second commit?